2010-02-08 19:03 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/gtwin/gtwin.c     * ULONG -> HB_SIZE
  * src/rtl/gtwvt/gtwvt.c
    * LONG -> int
    * ULONG -> HB_SIZE

  * contrib/gtwvg/gtwvg.c
  * contrib/gtwvg/wvggui.c
    * LONG -> int
    * ULONG -> HB_SIZE

  * contrib/gtwvg/wvgwin.c
    * ULONG -> HB_ULONG
    * LONG -> long

  * contrib/gtwvg/wvgutils.c
  * contrib/gtwvg/wvgsink.c
  * contrib/gtwvg/wvgwing.c
  * contrib/hbwin/axcore.c
    * LONG -> long

  * contrib/gtwvg/wvgutils.c
    * "" -> NULL

+++ harbour/contrib/gtwvg/wvgwing.c
@@ -1013,7 +1013,7 @@ HB_FUNC( WVG_POINTSIZETOHEIGHT )
 {
    HDC hdc = HB_ISNUM( 1 ) ? wvg_parhdc( 1 ) : GetDC( GetDesktopWindow() );
 
-   hb_retnl( ( LONG ) -MulDiv( ( LONG ) hb_parnl( 2 ), GetDeviceCaps( hdc, LOGPIXELSY ), 72 ) );
+   hb_retnl( ( long ) -MulDiv( ( LONG ) hb_parnl( 2 ), GetDeviceCaps( hdc, LOGPIXELSY ), 72 ) );
 
    if( !HB_ISNUM( 1 ) )
       ReleaseDC( GetDesktopWindow(), hdc );
@@ -1027,7 +1027,7 @@ HB_FUNC( WVG_HEIGHTTOPOINTSIZE )
 {
    HDC hdc = HB_ISNUM( 1 ) ? wvg_parhdc( 1 ) : GetDC( GetDesktopWindow() );
 
-   hb_retnl( ( LONG ) -MulDiv( hb_parnl( 2 ), 72, GetDeviceCaps( hdc, LOGPIXELSY ) ) );
+   hb_retnl( ( long ) -MulDiv( ( LONG ) hb_parnl( 2 ), 72, GetDeviceCaps( hdc, LOGPIXELSY ) ) );
 
    if( !HB_ISNUM( 1 ) )
       ReleaseDC( GetDesktopWindow(), hdc );
